pcb, sch: replace CR123A battery with 18650
Turns out that the  CR123A LiPo 3.7V batteries don't have a big market and
limits out maximum battery capacity to around ~800mAh.

This required quite a big of a board re-design - schematic is 99% the
same. So no big functional breaks are to be expected.

We lost a bit of the beautiful symmetry as the heatsinks below the
BQ25185 are super important.

Also dropped one I2C port again as the space is needed.